producers/requirements.txt
kafka-python==2.3.1
requests==2.31.0
python-dotenv==1.2.2